test,conformance-ingress: Change sanity check to deal with ICMP reply
The current ingress conformance test adds an ingress and then does a
curl to confirm it works. In the past, Cilium would have dropped the
request packet until the datapath was setup. This silent dropping causes
curl to retry for 15 seconds before giving up.

With the ICMP reply however, curl gets an immediate response and will
give up immediately. This causes the test to fail. So this commit
adds manual retry logic and delays in the test script.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>